+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package ess_imu_driver2
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.4 (2026-07-20)
+------------------
+* Moved sensor sampling & publishing to background thread, added G355 support
+
+2.0.3 (2024-11-29)
+------------------
+* bugfix - TimeCorrection class for PPS input
+
 2.0.2 (2024-10-25)
 ------------------
 * Add changes stating support for Jazzy